requirements.txt
discord.py==2.6.4
aiohttp==3.13.3